+++ ui/src/components/Sidebar.tsx
@@ -4,7 +4,6 @@ import {
   Target,
   LayoutDashboard,
   DollarSign,
-  Gauge,
   History,
   Search,
   SquarePen,
@@ -99,7 +98,6 @@ export function Sidebar() {
         <SidebarSection label="Company">
           <SidebarNavItem to="/org" label="Org" icon={Network} />
           <SidebarNavItem to="/costs" label="Costs" icon={DollarSign} />
-          <SidebarNavItem to="/usage" label="Usage" icon={Gauge} />
           <SidebarNavItem to="/activity" label="Activity" icon={History} />
           <SidebarNavItem to="/company/settings" label="Settings" icon={Settings} />
         </SidebarSection>
